fix some module naming and file location issues preventing auto-loading.
